third_party/WebKit/LayoutTests/http/tests/inspector/resource-tree/resource-tree-non-unique-url.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../resources-test.js"></script>
<script src="resource-tree-test.js"></script>
<script>
function loadIframe()
{
    var iframe = document.createElement("iframe");
    document.body.appendChild(iframe);
    iframe.src = "resources/resource-tree-non-unique-url-iframe.html";
}

function test() {
  TestRunner.addSniffer(Resources.FrameTreeElement.prototype, 'appendResource', onResource, true);

  var cssRequestsCount = 0;
  function onResource(resource) {
    if (resource.url.match(/\.css$/) && ++cssRequestsCount === 2) {
      TestRunner.addResult('Resources Tree:');
      ApplicationTestRunner.dumpResourcesTree();
      TestRunner.completeTest();
    }
  }

  TestRunner.evaluateInPage('loadIframe();');
}
</script>
</head>
<body onload="runTest()">
<p>Tests resources panel shows several resources with the same url if they were loaded with inspector already opened.</p>
</body>
</html>
