third_party/WebKit/LayoutTests/http/tests/inspector/resource-tree/resource-tree-non-unique-url-expected.txt

Tests resources panel shows several resources with the same url if they were loaded with inspector already opened.


Resources Tree:
Frames
    (resource-tree-non-unique-url.html)
        (resource-tree-non-unique-url-iframe.html)
            Images
                styles-non-unique-url.css
            Style
                styles-non-unique-url.css
            resource-tree-non-unique-url-iframe.html
        Script
            inspector-test.js
            resource-tree-test.js
            resources-test.js
        resource-tree-non-unique-url.html
